general: * local vars now work: ``` ❯ ./a.out > a: 10 10 > b: 20 20 > c: a + b 30 > c 30 > ```
app.c:
* Compiler needs interns for logging.

const.h:
* ConstId should be u16, since that's what we emit.

log.h:
* Added: CMP_BADREF, CMP_BADREFTYPE.

compiler.c/h:
* Compiler takes interns for logging.
* Resetting the compiler for REPL now sets reg to the number of in-use registers, so we can reuse them over repl lines. This will change when we add functions, since funcs are stored separate and dont consume register slots.

func.h:
* Changed FuncId to u16 since we will eventually use them for OP_CALL emits.

sym.c/h:
* SymGet returns a sym pointer now. Cleans up call sites. We dont cache pointers, so we dont care about invalidation.

+++ src/base/const/const.h
@@ -5,7 +5,7 @@
 #include "../cfg.h"
 #include "../intern/intern.h"
 
-typedef u32 ConstId;
+typedef u16 ConstId;
 
 typedef enum ConstType {
 	CONST_NUM,
